refactor: Se coloca el modal de la venta en un componente aparte en la carpera components/modals
